refactor: rename tryCatch utility and swagger setup files for consistency; add response handler utility
